+++ fsic/core.py
@@ -583,6 +583,90 @@ class VectorContainer:
         index_re = re.compile(r'\[\s*(.+?)?\s*\]')
         return index_re.sub(resolve_indexes, expression)
 
+    def reindex(
+            self, span: Sequence[Hashable], *, fill_value: Any = None, **fill_values: Any
+    ) -> 'VectorContainer':
+        """Return a copy of the current object, adjusted to match `span`. Values in overlapping periods between the old and new objects are preserved (copied over).
+
+        TODO: Consider implementing something closer to `pandas` `reindex()`
+              behaviour?
+
+        Parameters
+        ----------
+        span : iterable
+            Sequence of periods defining the span of the object to be returned
+        fill_value :
+            Default fill value for new periods
+        **fill_values :
+            Variable-specific fill value(s)
+        """
+        # Construct a mapping of:
+        #  - keys: positions in `span` (the new, reindexed object)
+        #  - values: the corresponding positions in `self.span` (i.e. where to
+        #            get the old values from)
+        positions = {}
+        for i, period in enumerate(span):
+            if period in self.span:
+                positions[i] = self._locate_period_in_span(period)
+
+        # Copy the current object and adjust:
+        #  - the span
+        #  - the individual underlying variables to conform to the new span
+        reindexed = self.copy()
+        reindexed.__dict__['span'] = span  # Use to bypass `strict`
+
+        for name in reindexed.index:
+            # Replace the underlying variable, to bypass dimension and type
+            # checks
+
+            # TODO: How to improve performance here?
+
+            # Get the fill value either as:
+            #  - specified (as a keyword argument in **fill_values)
+            #  - the default in `fill_value`
+            value = fill_values.get(name, fill_value)
+
+            # Special handling for `bool`, `int` and `str`-like dtypes (none of
+            # which have support for NaN/None)
+            if np.issubdtype(self[name].dtype, bool):
+                if value is None:
+                    value = False
+                else:
+                    value = bool(value)
+
+            elif np.issubdtype(self[name].dtype, np.integer):
+                if value is None:
+                    value = 0
+                else:
+                    value = int(value)
+
+            elif np.issubdtype(self[name].dtype, str):
+                if value is None:
+                    value = ''
+                else:
+                    value = str(value)
+
+            # Initialise the replacement with the correct length, and the fill
+            # value
+            reindexed.__dict__[f'_{name}'] = np.full(
+                len(span), value, dtype=self[name].dtype
+            )
+
+            # Copy over individual values
+            # TODO: Vectorise this?
+            for new, old in positions.items():
+                reindexed[name][new] = self[name][old]
+
+        return reindexed
+
+    def to_dataframe(self) -> 'pandas.DataFrame':
+        """Return the contents of the container as a `pandas` DataFrame, one column per variable. **Requires `pandas`**."""
+        from pandas import DataFrame
+
+        # NB Take variables one at a time, rather than use `self.values`. This
+        #    preserves the dtypes of the individual series.
+        return DataFrame({k: self[k] for k in self.index}, index=self.span)
+
     def eval(
         self,
         expression: str,
@@ -1424,6 +1508,34 @@ class BaseModel(SolverMixin, ModelInterface):
 
         return cls(index, *args, **{k: v.values for k, v in data.items()}, **kwargs)
 
+    def reindex(
+            self, span: Sequence[Hashable], *, fill_value: Any = None, **fill_values: Any
+    ) -> 'BaseModel':
+        """Return a copy of the current object, adjusted to match `span`. Values in overlapping periods between the old and new objects are preserved (copied over).
+
+        Parameters
+        ----------
+        span : iterable
+            Sequence of periods defining the span of the object to be returned
+        fill_value :
+            Default fill value for new periods
+        **fill_values :
+            Variable-specific fill value(s)
+
+        Notes
+        -----
+        These attributes have the following defaults (which can be over-ridden
+        in the usual way, using `fill_values`):
+         - status: '-' (`SolutionStatus.UNSOLVED.value`)
+         - iterations: -1
+
+        TODO: Consider how to generalise defaults to support extensions
+        """
+        fill_values['status'] = fill_values.get('status', SolutionStatus.UNSOLVED.value)
+        fill_values['iterations'] = fill_values.get('iterations', -1)
+
+        return super().reindex(span, fill_value=fill_value, **fill_values)
+
     def to_dataframe(
         self, *, status: bool = True, iterations: bool = True
     ) -> 'pandas.DataFrame':  # noqa: F821
@@ -1883,6 +1995,13 @@ Spans of submodels differ:
 
     def __deepcopy__(self, *args, **kwargs) -> 'BaseLinker':
         return self.copy()
+
+    def reindex(
+            self, span: Sequence[Hashable], *args: Any, **kwargs: Any
+    ) -> 'BaseLinker':
+        # TODO: Still to consider design of the `BaseLinker` equivalent to the
+        #       (now implemented) `VectorContainer` and `BaseModel` versions
+        raise NotImplementedError('`reindex()` method not yet implemented in `BaseLinker`')
 
     def to_dataframe(
         self, *, status: bool = True, iterations: bool = True